+++ src/components/dashboard/edit/BebridgeContainer.tsx
@@ -1,7 +1,7 @@
 import { ChangeEvent, useEffect, useState } from "react";
 import styled from "styled-components";
 import styles from "./EditPage.style.module.scss";
-import { useEdit } from "@/src/contexts/EditDashbordProvider";
+import { useEdit } from "@/src/contexts/EditDashboardProvider";
 import { Button } from "../../button/CustomButton2";
 import IconCheck from "@/public/images/dashboard/edit/ic_check.svg";
 
@@ -37,13 +37,12 @@ export default function BebridgeContainer() {
   useEffect(() => {
     if (
       isUpdateTitle !== "" &&
-      (isTitle !== isUpdateTitle ||
-      isColor !== isUpdateColor)
+      (isTitle !== isUpdateTitle || isColor !== isUpdateColor)
     ) {
       setIsDisabled(false);
     } else if (
       isUpdateTitle === "" ||
-      isTitle === isUpdateTitle && isColor === isUpdateColor
+      (isTitle === isUpdateTitle && isColor === isUpdateColor)
     ) {
       setIsDisabled(true);
     }
